+++ libs/gltfio/src/FFilamentAsset.h
@@ -40,6 +40,7 @@
 #include "Wireframe.h"
 #include "DependencyGraph.h"
 #include "DracoCache.h"
+#include "FFilamentInstance.h"
 
 #include <tsl/robin_map.h>
 
@@ -48,13 +49,6 @@
 
 namespace gltfio {
 namespace details {
-
-struct Skin {
-    std::string name;
-    std::vector<filament::math::mat4f> inverseBindMatrices;
-    std::vector<utils::Entity> joints;
-    std::vector<utils::Entity> targets;
-};
 
 // Encapsulates VertexBuffer::setBufferAt() or IndexBuffer::setBuffer().
 struct BufferSlot {
@@ -81,6 +75,14 @@ struct FFilamentAsset : public FilamentAsset {
 
     ~FFilamentAsset() {
         releaseSourceData();
+
+        // The only things we need to free in the instances are their animators.
+        // The union of all instance entities will be destroyed below.
+        for (FFilamentInstance* instance : mInstances) {
+            delete instance->animator;
+            delete instance;
+        }
+
         delete mAnimator;
         delete mWireframe;
         mEngine->destroy(mRoot);
@@ -159,7 +161,7 @@ struct FFilamentAsset : public FilamentAsset {
 
     Animator* getAnimator() noexcept {
         if (!mAnimator) {
-            mAnimator = new Animator(this);
+            mAnimator = new Animator(this, nullptr);
         }
         return mAnimator;
     }
@@ -185,10 +187,21 @@ struct FFilamentAsset : public FilamentAsset {
         mBufferSlots = {};
         mTextureSlots = {};
         releaseSourceAsset();
+        for (FFilamentInstance* instance : mInstances) {
+            instance->nodeMap = {};
+        }
     }
 
     const void* getSourceAsset() noexcept {
         return mSourceAsset;
+    }
+
+    FilamentInstance** getAssetInstances() noexcept {
+        return (FilamentInstance**) mInstances.data();
+    }
+
+    size_t getAssetInstanceCount() const noexcept {
+        return mInstances.size();
     }
 
     void acquireSourceAsset() {
@@ -230,7 +243,8 @@ struct FFilamentAsset : public FilamentAsset {
     std::vector<filament::Texture*> mTextures;
     filament::Aabb mBoundingBox;
     utils::Entity mRoot;
-    std::vector<Skin> mSkins;
+    std::vector<FFilamentInstance*> mInstances;
+    SkinVector mSkins; // unused for instanced assets
     Animator* mAnimator = nullptr;
     Wireframe* mWireframe = nullptr;
     int mSourceAssetRefCount = 0;
@@ -248,7 +262,7 @@ struct FFilamentAsset : public FilamentAsset {
     std::vector<TextureSlot> mTextureSlots;
     std::vector<const char*> mResourceUris;
     const cgltf_data* mSourceAsset = nullptr;
-    tsl::robin_map<const cgltf_node*, utils::Entity> mNodeMap;
+    NodeMap mNodeMap; // unused for instanced assets
     std::vector<std::pair<const cgltf_primitive*, filament::VertexBuffer*> > mPrimitives;
 };